requirements.txt
torch>=2.4.0
chromadb>=0.4.22
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
transformers>=4.40.0
sentence-transformers>=2.2.2
open_clip_torch>=2.24.0
openai>=1.0.0
requests>=2.28.0
